--- a/tests/test_validation.py
+++ b/tests/test_validation.py
@@ -1,9 +1,10 @@
 # coding: utf-8
 
-from cleanlab.internal import validation
 import numpy as np
 import pandas as pd
 import pytest
+
+from cleanlab.internal import validation
 
 
 @pytest.mark.parametrize("y_list", [["a", "b", "a"], [0, 1, 2]])
@@ -27,3 +28,9 @@ def test_label_to_array_raises_error():
     y = pd.DataFrame({"a": [0, 1], "b": [2, 3]})
     with pytest.raises(ValueError):
         validation.labels_to_array(y)
+
+
+def test_assert_valid_class_labels_fails_with_str_labels():
+    str_labels = np.array(["b", "b", "a", "c", "a"])
+    with pytest.raises(ValueError, match="string"):
+        validation.assert_valid_class_labels(str_labels)